backend/requirements.txt
Flask==2.3.2
flask-cors==4.0.0
Pillow==10.0.1
numpy==1.26.0
requests==2.31.0
APScheduler==3.10.1
python-dotenv==1.0.0
